requirements.txt
digikey-api
python-dotenv
dotenv